Moved storage model classes to its own file.
